fastapi-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
python-dotenv==1.0.1

